added the beginning of an integration test, starting with the addition of a destroydb method.
also, I've come up w/ a set of conventions for interacting with
the database.  these are meant to address a recurring problem
I've had using SQLAlchemy: existing references to db-backed
objects get stale when you do a commit, but there's no obvious
way to indicate which methods call .commit() and which don't.

I'll address this futher in the readme.md once I get to it ...